Changed initialisation to processCharts
- Changed initialisation to processCharts.
- Refactored chart processing.
