+++ src/init/index.js
@@ -1,7 +1,7 @@
 require("./glimma");
 require("./init");
 
-require("./initialise");
+require("./process_charts");
 require("./process_linkage");
 require("./process_inputs");
 require("./process_tables");
